アプリケーション作成部分をクラス化
